// generated/src/app/api-models/final-non-ap-setup-sel.model.ts
/* Copyright (c) 2020 . All Rights Reserved. */

export class FinalNonApSetupSel {
  seqFinalId: number;
  seqFinalIdSel: number;
  columnName: string;
  columnOccurrence: number;
  operator: string;
  fromValue: string;
  thruValue: string;
  securityCode: string;
  insertDatetime: Date;
  insertUser: string;
  insertProcess: string;
  updateDatetime: Date;
  updateUser: string;
  updateProcess: string;
  columnType: string;
}